automated_detection_testing/ci/detection_testing_batch/requirements.txt
ansible==2.9.20
ansible-runner==1.4.4
attackcti==0.3.4.3
boto3==1.11.0
GitPython==3.1.14
Jinja2==3.0.0
PyYAML==5.4
requests==2.25.1
six==1.16.0